Simplify StatsManager by delegating stats operations to new a class, ViewManager.
Classes in the StatsManager class hierarchy should just specify the classes
that are used for different stats implementations.  This commit moves all
stats-handling logic out of StatsManagerImplBase into a new class, ViewManager.
